consolidate some tests
